requirements.txt
Flask
line-bot-sdk==3.*
psycopg2-binary
python-dotenv
